src/requirements.txt
stemming
pystemmer
